Delete completed one-time scripts; move ops tooling to scripts/ops/
Deleted (git history is the archive):
- Debug scraps: test-debug*.mjs, test-poly-debug2.ts, test-polycentrism-debug.*
- Version-clamp project (ran Apr, re-applied Jul): apply/backfill/dry-run
  version-tag-clamp, backfill-map-version-clamp, record-dry-run-attribution,
  reapply-map-version-clamp-raw
- Completed one-time seeds/backfills: backfill-content-announcements,
  seed-included-cities, convert-galleries-to-webp

Moved to scripts/ops/ (manual recovery/diagnostic tools with ongoing value;
pnpm script names unchanged): audit-download-history,
rebuild-download-history-from-git, rebuild-download-version-buckets,
backfill-download-attribution, spotcheck-attribution-logs,
create-manual-download-attribution, audit-shared-map-attribution (+ .sh),
backfill-website-analytics, backfill-data-quality.

resolveRepoRoot now treats ops/ like dist/ (one level deeper than scripts/).
