+++ scripts/package.json
@@ -17,7 +17,7 @@
     "update-listing": "tsx update-listing.ts",
     "update-author": "tsx update-author.ts",
     "apply-data-quality": "tsx apply-data-quality.ts",
-    "backfill-data-quality": "tsx backfill-data-quality.ts",
+    "backfill-data-quality": "tsx ops/backfill-data-quality.ts",
     "check-data-quality": "tsx check-data-quality.ts",
     "format-manifests": "tsx format-manifests.ts",
     "regenerate-indexes": "tsx regenerate-indexes.ts",
@@ -31,20 +31,19 @@
     "generate-downloads:maps:download-only": "tsx generate-downloads.ts --type map --mode download-only",
     "generate-downloads:mods:download-only": "tsx generate-downloads.ts --type mod --mode download-only",
     "generate-download-history": "tsx generate-download-history.ts",
-    "audit-download-history": "tsx audit-download-history.ts",
-    "backfill-content-announcements": "tsx backfill-content-announcements.ts",
-    "audit-shared-map-attribution": "tsx audit-shared-map-attribution.ts",
-    "rebuild-download-history-from-git": "tsx rebuild-download-history-from-git.ts",
+    "audit-download-history": "tsx ops/audit-download-history.ts",
+    "audit-shared-map-attribution": "tsx ops/audit-shared-map-attribution.ts",
+    "rebuild-download-history-from-git": "tsx ops/rebuild-download-history-from-git.ts",
     "merge-download-attribution": "tsx merge-download-attribution.ts",
-    "backfill-download-attribution": "tsx backfill-download-attribution.ts",
-    "spotcheck-attribution-logs": "tsx spotcheck-attribution-logs.ts",
-    "create-manual-download-attribution": "tsx create-manual-download-attribution.ts",
-    "rebuild-download-version-buckets": "tsx rebuild-download-version-buckets.ts",
+    "backfill-download-attribution": "tsx ops/backfill-download-attribution.ts",
+    "spotcheck-attribution-logs": "tsx ops/spotcheck-attribution-logs.ts",
+    "create-manual-download-attribution": "tsx ops/create-manual-download-attribution.ts",
+    "rebuild-download-version-buckets": "tsx ops/rebuild-download-version-buckets.ts",
     "reconcile-attributed-downloads": "tsx reconcile-attributed-downloads.ts",
     "analytics": "tsx generate-analytics.ts",
     "capture-railyard-app-downloads": "tsx capture-railyard-app-downloads.ts",
     "capture-website-analytics": "tsx capture-website-analytics.ts",
-    "backfill-website-analytics": "tsx backfill-website-analytics.ts --days 7",
+    "backfill-website-analytics": "tsx ops/backfill-website-analytics.ts --days 7",
     "generate-website-analytics": "tsx generate-website-analytics.ts",
     "backfill-discord-server-metrics": "tsx capture-discord-server-metrics.ts --reset-history",
     "capture-discord-server-metrics": "tsx capture-discord-server-metrics.ts",
@@ -61,7 +60,6 @@
     "generate-local-grid-statistics": "tsx generate-local-grid-statistics.ts",
     "generate-local-playable-area-debug": "tsx generate-local-playable-area-debug.ts",
     "generate-search-aliases": "tsx generate-search-aliases.ts",
-    "seed-included-cities": "tsx seed-included-cities.ts",
     "sync-map-file-sizes": "tsx sync-map-file-sizes.ts",
     "sync-last-updated": "tsx sync-last-updated.ts",
     "notify-discord": "tsx notify-discord.ts",
