+++ scripts/tsconfig.json
@@ -11,6 +11,6 @@
     "rootDir": ".",
     "resolveJsonModule": true
   },
-  "include": ["*.ts", "lib/**/*.ts", "tests/**/*.ts"],
+  "include": ["*.ts", "lib/**/*.ts", "ops/**/*.ts", "tests/**/*.ts"],
   "exclude": ["dist"]
 }
